+++ src/lib/themes/ThemeConfig.ts
@@ -22,6 +22,7 @@ export type ThemeConfig = {
 		default: ComponentType;
 		label: ComponentType;
 		number: ComponentType;
+		object: ComponentType;
 		tabs: ComponentType;
 		textarea: ComponentType;
 		text: ComponentType;
@@ -35,6 +36,7 @@ export type ThemeConfig = {
 		date: ComponentType;
 		default: ComponentType;
 		number: ComponentType;
+		object: ComponentType;
 		tabs: ComponentType;
 		textarea: ComponentType;
 		text: ComponentType;