+++ src/lib/themes/carbon/index.js
@@ -34,6 +34,7 @@ import ColumnsViewField from './ViewFieldsComponents/ColumnsField.svelte';
 import DateViewField from './ViewFieldsComponents/DateField.svelte';
 import DefaultViewField from './ViewFieldsComponents/DefaultField.svelte';
 import NumberViewField from './ViewFieldsComponents/NumberField.svelte';
+import ObjectViewField from './ViewFieldsComponents/ObjectField.svelte';
 import TabsViewField from './ViewFieldsComponents/TabsField.svelte';
 import TextareaViewField from './ViewFieldsComponents/DefaultField.svelte';
 import TextViewField from './ViewFieldsComponents/DefaultField.svelte';
@@ -71,6 +72,7 @@ const theme = {
 		default: DefaultViewField,
 		label: ViewLabel,
 		number: NumberViewField,
+		object: ObjectViewField,
 		tabs: TabsViewField,
 		text: TextViewField,
 		textarea: TextareaViewField,
@@ -83,6 +85,7 @@ const theme = {
 		date: DateFormField,
 		default: DefaultFormField,
 		number: NumberFormField,
+		object: TextFormField, // FIXME
 		tabs: TabsFormField,
 		text: TextFormField,
 		textarea: TextareaFormField,