+++ src/testApp/internal/testsInternal.ts
@@ -10,8 +10,18 @@ export type Test = {
 	url_field: string;
 	path_field: string;
 	date_field: Date;
+	object_field: object;
 	[key: string]: unknown;
 };
+
+function fakeObject(): object|string {
+	return {
+		data1: faker.lorem.words(faker.number.int({ min: 1, max: 4 })),
+		data2: {
+			nested: faker.lorem.words(faker.number.int({ min: 1, max: 4 })),
+		},
+	};
+}
 
 const baseTests: Array<Test> = Array(10)
 	.fill(undefined)
@@ -28,7 +38,8 @@ const baseTests: Array<Test> = Array(10)
 			toggle_field: faker.datatype.boolean(),
 			url_field: faker.internet.url(),
 			path_field: faker.system.filePath().replace(/^\/[^/]+\//g, '/'),
-			date_field: date
+			date_field: date,
+			object_field: fakeObject() as object
 		};
 	});